requirements.txt
playwright==1.46.0
flask==2.2.5
apscheduler==3.10.4
python-dotenv==1.0.0
requests
